Issue #412: Move Cursor files into the same pkg: browser.cursor.
