+++ app/src/main/java/org/mozilla/focus/browser/cursor/CursorView.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.widget
+package org.mozilla.focus.browser.cursor
 
 import android.content.Context
 import android.graphics.Canvas
@@ -14,8 +14,6 @@ import android.support.annotation.UiThread
 import android.util.AttributeSet
 import android.view.View
 import org.mozilla.focus.R
-import org.mozilla.focus.utils.Direction
-import java.util.HashSet
 import java.util.concurrent.TimeUnit
 
 private const val CURSOR_SIZE = 45f
