+++ app/src/main/java/org/mozilla/focus/browser/cursor/CursorViewModel.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.browser
+package org.mozilla.focus.browser.cursor
 
 import android.graphics.PointF
 import android.os.SystemClock
